+++ scglobals/static/tx/src/app/services/db.js
@@ -11,7 +11,7 @@ export default class DB {
     this.db.on('error', function (err) { console.log(err) });
     this.initialQueries = new InitialQueries(this.db);
   }
-  destroy_db(){
+  destroy(){
     return this.db.destroy();
   }
   initdb(){
@@ -23,10 +23,24 @@ export default class DB {
     return this.db.bulkDocs(docs);
   }
   query(q,ops){
-    ops = _.extend(
-      {reduce:false,limit:1000,include_docs:true},ops);
+    // putting here instead of in constructor as _.extend modifies actual object
+    var default_options = {reduce:false,limit:1000,include_docs:true};
+    if (!ops){
+      ops = default_options;
+    }
+    else if (!ops['reduce']){
+      ops = _.extend(ops,default_options);
+    }
     return this.db.query(q,ops).then((result)=> {
-      return _.pluck(result.rows, 'doc');
+      // making rows straight up docs instead of {_id,_rev,doc} bs
+      if (!ops['reduce']){
+        result['doc_rows'] = _.pluck(result.rows, 'doc');
+        delete result['rows'];
+      }
+      else{
+        return result.rows[0].value;
+      }
+      return result;
     }).catch(function (err) {
       return err;
     });