+++ scglobals/static/tx/changelog.md
@@ -3,8 +3,9 @@
 ## 2015-12-?
 * docs stored in couchdb
 User
-* sees 100 shipments as date, from, to, total items, total value | Sm | Sm
-* can click a shipment and see the shipment's transactions | already have the shipment loaded on list view, so this could be one less hit to the DB... | sm
+* sees 1000 shipments as date, from, to, total items, total value | Sm | 2h
+* can click a shipment and see the shipment's transactions | already have the shipment loaded on list view, so this could be one less hit to the DB... | sm | 1h
+* sees the total number of shipments, their total value, and their total number of items | Sm | 2h
 
 #### improvements